Restructured file conversion and removal
Restructured the file conversion using the las2las utility so that you only attempt to remove the LAS output file if and after the LAS file has been converted to a LAZ file.

Previously, the code would attempt to remove the LAS file even if the compression using las2las had not occurred.